+++ connecteur/chorus-par-csv/action/ChorusParCSVException.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+class ChorusParCSVException extends Exception
+{
+}