+++ connecteur/cpp/action/CPPListeFacture.php
@@ -1,45 +1,42 @@
 <?php
 
+declare(strict_types=1);
+
 class CPPListeFacture extends ActionExecutor
 {
     /**
-     * @return UTF8Encoder
-     */
-    public function getUTF8Encoder()
-    {
-        return $this->objectInstancier->getInstance(UTF8Encoder::class);
-    }
-
-    /**
-     * @return array|mixed
      * @throws Exception
      */
-    public function metier()
+    public function metier(): string
     {
         /** @var CPP $cpp */
         $cpp = $this->getMyConnecteur();
-        return $this->getUTF8Encoder()->decode(
-            json_encode($this->getUTF8Encoder()->encode(
-                $cpp->rechercheFactureParRecipiendaire("", $cpp->getDateDepuisLe(), $cpp->getDateJusquAu())
-            ))
+        return json_encode(
+            $cpp->rechercheFactureParRecipiendaire('', $cpp->getDateDepuisLe(), $cpp->getDateJusquAu()),
+            JSON_THROW_ON_ERROR
         );
     }
 
     /**
-     * @return bool
      * @throws Exception
      */
-    public function go()
+    public function go(): bool
     {
         /** @var CPP $cpp */
         $cpp = $this->getMyConnecteur();
         $result = $this->metier();
         if (! $result) {
-            $this->setLastMessage("La connexion cpp a échoué : " . $cpp->getLastError());
+            $this->setLastMessage('La connexion cpp a échoué : ' . $cpp->getLastError());
             return false;
         }
-        $this->setLastMessage("Liste des factures ayant changé de statut entre le " . $cpp->getDateDepuisLe() .
-            " et le " . $cpp->getDateJusquAu() . ": " . $result);
+        $this->setLastMessage(
+            sprintf(
+                'Liste des factures ayant changé de statut entre le %s et le %s : %s',
+                $cpp->getDateDepuisLe(),
+                $cpp->getDateJusquAu(),
+                $result
+            )
+        );
         return true;
     }
 }